Make infobubble an own component
